Added run-time support for AVX2 detection in bitshuffle too
