+++ blosc/CMakeLists.txt
@@ -37,16 +37,16 @@ if(NOT DEACTIVATE_ZLIB)
 endif(NOT DEACTIVATE_ZLIB)
 
 # library sources
-set(SOURCES blosc.c blosclz.c shuffle-generic.c)
+set(SOURCES blosc.c blosclz.c shuffle-generic.c bitshuffle-generic.c)
 if(COMPILER_SUPPORT_SSE2)
-    message(STATUS "Adding run-time support for SSE2.")
-    set(SOURCES ${SOURCES} shuffle-sse2.c)
+    message(STATUS "Adding run-time support for SSE2")
+    set(SOURCES ${SOURCES} shuffle-sse2.c bitshuffle-sse2.c)
 endif(COMPILER_SUPPORT_SSE2)
 if(COMPILER_SUPPORT_AVX2)
-    message(STATUS "Adding run-time support for AVX2.")
-    set(SOURCES ${SOURCES} shuffle-avx2.c)
+    message(STATUS "Adding run-time support for AVX2")
+    set(SOURCES ${SOURCES} shuffle-avx2.c bitshuffle-avx2.c)
 endif(COMPILER_SUPPORT_AVX2)
-set(SOURCES ${SOURCES} shuffle.c bitshuffle.c)
+set(SOURCES ${SOURCES} shuffle.c)
 
 # library install directory
 set(lib_dir lib${LIB_SUFFIX})
@@ -112,10 +112,10 @@ if(COMPILER_SUPPORT_SSE2)
     if (MSVC)
         # MSVC targets SSE2 by default on 64-bit configurations, but not 32-bit configurations.
         if (${CMAKE_SIZEOF_VOID_P} EQUAL 4)
-            set_source_files_properties(shuffle-sse2.c PROPERTIES COMPILE_FLAGS "/arch:SSE2")
+            set_source_files_properties(shuffle-sse2.c bitshuffle-sse2.c PROPERTIES COMPILE_FLAGS "/arch:SSE2")
         endif (${CMAKE_SIZEOF_VOID_P} EQUAL 4)
     else (MSVC)
-        set_source_files_properties(shuffle-sse2.c PROPERTIES COMPILE_FLAGS -msse2)
+        set_source_files_properties(shuffle-sse2.c bitshuffle-sse2.c PROPERTIES COMPILE_FLAGS -msse2)
     endif (MSVC)
 
     # Define a symbol for the shuffle-dispatch implementation
@@ -127,9 +127,9 @@ if(COMPILER_SUPPORT_SSE2)
 endif(COMPILER_SUPPORT_SSE2)
 if(COMPILER_SUPPORT_AVX2)
     if (MSVC)
-        set_source_files_properties(shuffle-avx2.c PROPERTIES COMPILE_FLAGS "/arch:AVX2")
+        set_source_files_properties(shuffle-avx2.c bitshuffle-avx2.c PROPERTIES COMPILE_FLAGS "/arch:AVX2")
     else (MSVC)
-        set_source_files_properties(shuffle-avx2.c PROPERTIES COMPILE_FLAGS -mavx2)
+        set_source_files_properties(shuffle-avx2.c bitshuffle-avx2.c PROPERTIES COMPILE_FLAGS -mavx2)
     endif (MSVC)
 
     # Define a symbol for the shuffle-dispatch implementation
