+++ blosc/shuffle.h
@@ -11,7 +11,7 @@
     Consumers should almost always prefer to call these routines instead
     of directly calling one of the hardware-accelerated routines, since
     these are cross-platform and future-proof. */
-    
+
 #ifndef SHUFFLE_H
 #define SHUFFLE_H
 
@@ -32,7 +32,7 @@ extern "C" {
   platform and future-proof.
 */
 BLOSC_NO_EXPORT void shuffle(const size_t bytesoftype, const size_t blocksize,
-                              const uint8_t* const _src, uint8_t* const _dest);
+                             const uint8_t* _src, const uint8_t* _dest);
 
 /**
   Primary unshuffle routine.
@@ -45,7 +45,7 @@ BLOSC_NO_EXPORT void shuffle(const size_t bytesoftype, const size_t blocksize,
   platform and future-proof.
 */
 BLOSC_NO_EXPORT void unshuffle(const size_t bytesoftype, const size_t blocksize,
-                                const uint8_t* const _src, uint8_t* const _dest);
+                               const uint8_t* _src, const uint8_t* _dest);
 
 #ifdef __cplusplus
 }
